Update Observer to pull rather than receive pushed data
This will allow our observers to fetch and consume only data that they
need rather than being handed the entire dataset by the Subject.

This is an optimization that will enable easy adding of more state
variables in the Subject later on without modifying the Observer
interface and all of its concrete implementations.